feat(native): support loading ApplicationData from manifest for FileSystemClassScanner
