+++ src/main/kotlin/de/idealo/security/endpointexporter/classreading/ClassScanner.kt
@@ -5,6 +5,7 @@ import de.idealo.security.endpointexporter.classreading.type.ClassMetadata
 import org.springframework.core.io.ResourceLoader
 import org.springframework.core.io.support.PathMatchingResourcePatternResolver
 import java.nio.file.Path
+import java.util.Properties
 import java.util.regex.Pattern
 
 abstract class ClassScanner(
@@ -17,7 +18,7 @@ abstract class ClassScanner(
 
     abstract fun scan(entrypoint: Path): List<ClassMetadata>
 
-    abstract fun scanApplicationData(entrypoint: Path): ApplicationMetadata?
+    abstract fun scanApplicationData(entrypoint: Path): ApplicationMetadata
 
     protected fun isCandidate(classMetadata: ClassMetadata): Boolean {
 
@@ -33,5 +34,27 @@ abstract class ClassScanner(
 
         // if no filter matches, exclude per default
         return false
+    }
+
+    protected fun loadApplicationDataFromManifest(resourcePattern: String): ApplicationMetadata {
+
+        val resource = resourcePatternResolver.getResource(resourcePattern)
+
+        return when {
+            resource.exists() -> {
+                val properties = resource.inputStream.use { stream ->
+                    Properties().also { it.load(stream) }
+                }
+                ApplicationMetadata(
+                    title = properties.getProperty("Implementation-Title") ?: "n/a",
+                    version = properties.getProperty("Implementation-Version") ?: "n/a",
+                )
+            }
+
+            else -> ApplicationMetadata(
+                title = "n/a",
+                version = "n/a"
+            )
+        }
     }
 }
